# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Playero API",
        "description": "API oficial do sistema Playero",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Autentica e retorna um token de acesso",
                "operationId": "a7997e7c1e3bfed64d4e56de82415931",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@playero.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado"
                    },
                    "401": {
                        "description": "Credenciais invalidas"
                    },
                    "403": {
                        "description": "Usuario inativo"
                    },
                    "422": {
                        "description": "Dados invalidos"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "f6b8957b4c0b25636613a9f0512a8ba9",
                "responses": {
                    "200": {
                        "description": "Logout realizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Usuario autenticado",
                "operationId": "37d33f0281f67d31d05b4c45eea0143b",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Renovar token",
                "operationId": "c42c3cd7fd7fe0dbd5bcc42962059e90",
                "responses": {
                    "200": {
                        "description": "Token renovado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/quadras": {
            "get": {
                "tags": [
                    "Quadras"
                ],
                "summary": "Listar quadras ativas",
                "operationId": "55fe97b2dc48e4feaf26788bf9ea4598",
                "responses": {
                    "200": {
                        "description": "Lista de quadras ativas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Quadras"
                ],
                "summary": "Criar quadra",
                "operationId": "f7a275eacdb45ba1a8e13c136ac7be2a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "tipo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Quadra 1"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "beach_tennis"
                                    },
                                    "ativa": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "ordem": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "capacidade": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quadra criada"
                    },
                    "422": {
                        "description": "Dados invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/quadras/{id}": {
            "put": {
                "tags": [
                    "Quadras"
                ],
                "summary": "Editar quadra",
                "operationId": "3a1516e782debf5efa6ea25b9e1e29fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Quadra 1"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "beach_tennis"
                                    },
                                    "ordem": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "capacidade": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quadra atualizada"
                    },
                    "404": {
                        "description": "Quadra nao encontrada"
                    },
                    "422": {
                        "description": "Dados invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Quadras"
                ],
                "summary": "Excluir quadra",
                "operationId": "2993536b59fb998e168aae7015894626",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quadra removida"
                    },
                    "404": {
                        "description": "Quadra nao encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/quadras/{id}/status": {
            "patch": {
                "tags": [
                    "Quadras"
                ],
                "summary": "Ativar ou desativar quadra",
                "operationId": "9b1abe8f5ff7eac73f7d6064690c79d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ativa"
                                ],
                                "properties": {
                                    "ativa": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status atualizado"
                    },
                    "404": {
                        "description": "Quadra nao encontrada"
                    },
                    "422": {
                        "description": "Dados invalidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}